Flask_Project/requirements.txt
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
cycler==0.10.0
decorator==4.4.2
einops==0.2.0
filelock==3.0.12
Flask==1.1.2
gdown==3.11.0
gevent==21.1.2
greenlet==1.0.0
idna==2.10
imageio==2.9.0
imgaug==0.4.0
itsdangerous==1.1.0
Jinja2==2.11.3
kiwisolver==1.3.1
lmdb==1.0.0
MarkupSafe==1.1.1
matplotlib==3.4.1
networkx==2.5.1
numpy==1.20.2
opencv-python==4.5.1.48
Pillow==8.2.0
prefetch-generator==1.0.1
pycparser==2.20
pyparsing==2.4.7
PySocks==1.7.1
python-dateutil==2.8.1
PyWavelets==1.1.1
PyYAML==5.4.1
requests==2.25.1
scikit-image==0.18.1
scipy==1.6.2
Shapely==1.7.1
six==1.15.0
tifffile==2021.4.8
torch==1.8.1+cpu
torchaudio==0.8.1
torchvision==0.9.1+cpu
tqdm==4.60.0
typing-extensions==3.7.4.3
urllib3==1.26.4
vietocr==0.3.5
Werkzeug==1.0.1
zope.event==4.5.0
zope.interface==5.4.0
